# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Slicer API\n基于Kubernetes资源的切片管理系统API\n包括 切片管理 监控管理 性能保证 等功能",
        "title": "Slicer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:30001",
    "basePath": "/",
    "paths": {
...
        "/sla": {
            "get": {
                "description": "获取系统内全部SLA列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SLA"
                ],
                "summary": "获取所有SLA",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SLA"
                            }
                        }
                    },
                    "500": {
                        "description": "获取失败/响应编码失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "接收SLA对象并创建新资源\nslice id对应切片必须存在",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SLA"
                ],
                "summary": "创建SLA",
                "parameters": [
                    {
                        "description": "SLA对象",
                        "name": "sla",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SLA"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功返回SLA对象",
                        "schema": {
                            "$ref": "#/definitions/model.SLA"
                        }
                    },
                    "400": {
                        "description": "请求解码失败/SLA已存在/参数非法",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "切片不存在",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "存储失败/响应编码失败等服务器内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sla/{slaId}": {
            "get": {
                "description": "根据SLA ID获取资源详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SLA"
                ],
                "summary": "获取单个SLA",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SLA ID",
                        "name": "slaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/model.SLA"
                        }
                    },
                    "400": {
                        "description": "缺少SLA ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "SLA不存在",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "获取失败/响应编码失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "根据SLA ID更新资源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SLA"
                ],
                "summary": "更新SLA",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SLA ID",
                        "name": "slaId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新后的SLA对象",
                        "name": "sla",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SLA"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功返回对象",
                        "schema": {
                            "$ref": "#/definitions/model.SLA"
                        }
                    },
                    "400": {
                        "description": "缺少SLA ID/请求解码失败/参数非法",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "SLA不存在",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "更新失败/响应编码失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据SLA ID删除资源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SLA"
                ],
                "summary": "删除SLA",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SLA ID",
                        "name": "slaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "缺少SLA ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "SLA不存在",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.SLA": {
            "type": "object",
            "properties": {
                "availability": {
                    "description": "可用性",
                    "type": "number"
                },
                "down_bandwidth": {
                    "description": "下行带宽",
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "latency": {
                    "description": "延迟",
                    "type": "number"
                },
                "slice_id": {
                    "type": "string"
                },
                "up_bandwidth": {
                    "description": "上行带宽",
                    "type": "number"
                }
            }
        },
...
    }
}